+++ doc/reference.rst
@@ -769,6 +769,35 @@ invocation of ``get()``.
 
 :Thread Safety: ``get()`` can be called from multiple threads simultaneously.
 
+.. |template_keyword| replace:: ``template_keyword``
+.. _template_keyword:
+
+``template_keyword``
+--------------------
+
+This class template encapsulates a named template parameter.  Every type
+generated by the |BOOST_PARAMETER_TEMPLATE_KEYWORD| macro is a specialization
+of |template_keyword|.
+
+:Defined in: `boost/parameter/template_keyword.hpp`__
+
+__ ../../../../boost/parameter/template_keyword.hpp
+
+.. parsed-literal::
+
+    template <typename Tag, typename T>
+    struct template_keyword
+    {
+        typedef Tag key_type;
+        typedef T value_type;
+        typedef *implementation defined* reference;
+    };
+
+The |ntp_cpp|_ test program demonstrates proper usage of this class template.
+
+.. |ntp_cpp| replace:: ntp.cpp
+.. _ntp_cpp: ../../test/ntp.cpp
+
 ``parameters``
 --------------
 
@@ -784,6 +813,8 @@ corresponding template argument to ``parameters``.
 __ ../../../../boost/parameter/parameters.hpp
 
 **If** |BOOST_PARAMETER_HAS_PERFECT_FORWARDING| is ``#defined``, **then**
+
+.. parsed-literal::
 
     template <typename ...PSpec>
     struct parameters
@@ -2015,9 +2046,13 @@ Treats *name* as if it were of the form:
 ``BOOST_PARAMETER_TEMPLATE_KEYWORD(name)``
 ------------------------------------------
 
-:Defined in: `boost/parameter/name.hpp`__
+:Defined in: |template_keyword_header|_
+:Included by: |name_header|_
 
-__ ../../../../boost/parameter/name.hpp
+.. |template_keyword_header| replace:: boost/parameter/template_keyword.hpp
+.. _template_keyword_header: ../../../../boost/parameter/template_keyword.hpp
+.. |name_header| replace:: boost/parameter/name.hpp
+.. _name_header: ../../../../boost/parameter/name.hpp
 
 Expands to:
