+++ include/boost/parameter/template_keyword.hpp
@@ -1,0 +1,66 @@
+// Copyright Daniel Wallin 2006.
+// Copyright Cromwell D. Enage 2017.
+// Distributed under the Boost Software License, Version 1.0.
+// (See accompanying file LICENSE_1_0.txt or copy at
+// http://www.boost.org/LICENSE_1_0.txt)
+
+#ifndef BOOST_PARAMETER_TEMPLATE_KEYWORD_HPP
+#define BOOST_PARAMETER_TEMPLATE_KEYWORD_HPP
+
+#include <boost/parameter/aux_/template_keyword.hpp>
+#include <boost/mpl/bool.hpp>
+#include <boost/mpl/if.hpp>
+#include <boost/mpl/eval_if.hpp>
+#include <boost/mpl/identity.hpp>
+#include <boost/type_traits/add_lvalue_reference.hpp>
+#include <boost/type_traits/is_function.hpp>
+#include <boost/type_traits/is_array.hpp>
+
+namespace boost { namespace parameter { 
+
+    template <typename Tag, typename T>
+    struct template_keyword : ::boost::parameter::aux::template_keyword_base
+    {
+        typedef Tag key_type;
+        typedef T value_type;
+
+        // reference is needed for two reasons:
+        //
+        // 1. It is used in the body of arg_list<...>
+        //
+        // 2. It is the result of binding<...>, which we mistakenly told
+        //    people to use instead of value_type<...> to access named
+        //    template parameters
+        //
+        // It used to be that reference == value_type, but that broke when
+        // the argument was a function or array type, because various
+        // arg_list functions return reference.
+        //
+        // Simply making reference == value_type& would break all the
+        // legacy code that uses binding<...> to access named template
+        // parameters. -- David Abrahams
+        typedef typename ::boost::mpl::eval_if<
+            typename ::boost::mpl::if_<
+                ::boost::is_function<value_type>
+              , ::boost::mpl::true_
+              , ::boost::is_array<value_type>
+            >::type
+          , ::boost::add_lvalue_reference<value_type>
+          , ::boost::mpl::identity<value_type>
+        >::type reference;
+    };
+}} // namespace boost::parameter
+
+#define BOOST_PARAMETER_TEMPLATE_KEYWORD(name)                               \
+    namespace tag                                                            \
+    {                                                                        \
+        struct name;                                                         \
+    }                                                                        \
+    template <typename T>                                                    \
+    struct name : ::boost::parameter::template_keyword<tag::name,T>          \
+    {                                                                        \
+    };
+/**/
+
+#endif  // include guard
+
